More improvements for mobile.

+++ QuestomAssets/AssetOps/DeletePlaylistOp.cs
@@ -7,12 +7,14 @@ namespace QuestomAssets.AssetOps
 {
     public class DeletePlaylistOp : AssetOp
     {
-        public DeletePlaylistOp(string playlistID)
+        public DeletePlaylistOp(string playlistID, bool deleteSongsOnPlaylist = true)
         {
             PlaylistID = playlistID;
+            DeleteSongsOnPlaylist = deleteSongsOnPlaylist;
         }
 
         public string PlaylistID { get; private set; }
+        public bool DeleteSongsOnPlaylist { get; private set; }
         internal override void PerformOp(OpContext context)
         {
             if (string.IsNullOrEmpty(PlaylistID))
@@ -23,19 +25,21 @@ namespace QuestomAssets.AssetOps
 
             var playlist = context.Cache.PlaylistCache[PlaylistID];
 
-            try
+            if (DeleteSongsOnPlaylist)
             {
-                foreach (var song in playlist.Songs)
+                try
                 {
-                    OpCommon.DeleteSong(context, song.Key);
+                    foreach (var song in playlist.Songs.ToList())
+                    {
+                        OpCommon.DeleteSong(context, song.Key);
+                    }
                 }
+                catch (Exception ex)
+                {
+                    throw new Exception($"Error while deleting songs from playlist id {playlist.Playlist.PackID}", ex);
+                }
+                playlist.Songs.Clear();
             }
-            catch (Exception ex)
-            {
-                throw new Exception($"Error while deleting songs from playlist id {playlist.Playlist.PackID}", ex);
-            }
-            playlist.Songs.Clear();
-
             //this should be done in song delete already
             //playlist.Playlist.BeatmapLevelCollection.Object.BeatmapLevels.ForEach(x => { x.Target.ParentFile.DeleteObject(x.Object); x.Dispose(); });
             var mlp = context.Engine.GetMainLevelPack();
